fix:单元测试
- REST消息测试，注入顺序问题
